moved up fuzzy specs to the channel package
